Get toggle working and fix tests
