+++ clients/admin-ui/src/features/common/table/constants.ts
@@ -1,7 +1,13 @@
 export const MECHANISM_MAP = new Map([
-  ["opt_in", "Opt In"],
-  ["necessary", "Necessary"],
-  ["opt_out", "Opt Out"],
+  ["opt_in", "Opt in"],
+  ["notice_only", "Notice only"],
+  ["opt_out", "Opt out"],
+]);
+
+export const ENFORCEMENT_LEVEL_MAP = new Map([
+  ["system_wide", "System wide"],
+  ["frontend", "Front end"],
+  ["not_applicable", "Not applicable"],
 ]);
 
 export const FIELD_TYPE_MAP = new Map([
