+++ clients/admin-ui/src/features/common/table/index.ts
@@ -1,6 +1,5 @@
 export {
   DateCell,
-  EnableCell,
   MechanismCell,
   MultiTagCell,
   TitleCell,
