--- clients/admin-ui/src/features/privacy-notices/constants.ts
@@ -1,11 +1,0 @@
-export const MECHANISM_MAP = new Map([
-  ["opt_in", "Opt in"],
-  ["notice_only", "Notice only"],
-  ["opt_out", "Opt out"],
-]);
-
-export const ENFORCEMENT_LEVEL_MAP = new Map([
-  ["system_wide", "System wide"],
-  ["frontend", "Front end"],
-  ["not_applicable", "Not applicable"],
-]);
