+++ clients/admin-ui/src/pages/consent/privacy-notices/index.tsx
@@ -1,10 +1,20 @@
 import { PRIVACY_REQUESTS_ROUTE } from "@fidesui/components";
-import { Box, Breadcrumb, BreadcrumbItem, Heading, Text } from "@fidesui/react";
+import {
+  Box,
+  Breadcrumb,
+  BreadcrumbItem,
+  Heading,
+  Switch,
+  Text,
+  useDisclosure,
+  WarningIcon,
+} from "@fidesui/react";
+import ConfirmationModal from "common/ConfirmationModal";
 import { useHasPermission } from "common/Restrict";
 import EmptyTableState from "common/table/EmptyTableState";
 import NextLink from "next/link";
-import { useMemo } from "react";
-import { Column } from "react-table";
+import { ChangeEvent, useMemo } from "react";
+import { CellProps, Column } from "react-table";
 
 import { useAppSelector } from "~/app/hooks";
 import Layout from "~/features/common/Layout";
@@ -14,7 +24,6 @@ import {
 } from "~/features/common/nav/v2/routes";
 import {
   DateCell,
-  EnableCell,
   FidesTable,
   MechanismCell,
   MultiTagCell,
@@ -29,6 +38,71 @@ import {
   usePatchPrivacyNoticesMutation,
 } from "~/features/privacy-notices/privacy-notices.slice";
 import { PrivacyNoticeResponse, ScopeRegistryEnum } from "~/types/api";
+
+type EnableCellProps<T extends object> = CellProps<T, boolean> & {
+  onToggle: (data: any) => Promise<any>;
+};
+
+export const EnableCell = <T extends object>({
+  value,
+  column,
+  row,
+  onToggle,
+}: EnableCellProps<T>) => {
+  const modal = useDisclosure();
+  const handlePatch = async ({ enable }: { enable: boolean }) => {
+    // @ts-ignore
+    await onToggle([{ id: row.original.id, disabled: !enable }]);
+  };
+
+  const handleToggle = async (event: ChangeEvent<HTMLInputElement>) => {
+    const { checked } = event.target;
+    if (checked) {
+      await handlePatch({ enable: true });
+    } else {
+      modal.onOpen();
+    }
+  };
+
+  return (
+    <>
+      <Switch
+        colorScheme="complimentary"
+        isChecked={!value}
+        data-testid={`toggle-${column.Header}`}
+        /**
+         * It's difficult to use a custom column in react-table 7 since we'd have to modify
+         * the declaration file. However, that modifies the type globally, so our datamap table
+         * would also have issues. Ignoring the type for now, but should potentially revisit
+         * if we update to react-table 8
+         * https://github.com/DefinitelyTyped/DefinitelyTyped/discussions/59837
+         */
+        // @ts-ignore
+        disabled={column.disabled}
+        onChange={handleToggle}
+      />
+      <ConfirmationModal
+        isOpen={modal.isOpen}
+        onClose={modal.onClose}
+        onConfirm={() => {
+          handlePatch({ enable: false });
+          modal.onClose();
+        }}
+        title="Disable privacy notice"
+        message={
+          <Text color="gray.500">
+            Are you sure you want to disable this privacy notice? Disabling this
+            notice means your users will no longer see this explanation about
+            your data uses which is necessary to ensure compliance.
+          </Text>
+        }
+        continueButtonText="Confirm"
+        isCentered
+        icon={<WarningIcon color="orange.100" />}
+      />
+    </>
+  );
+};
 
 const PrivacyNoticesPage = () => {
   // Subscribe to get all privacy notices
