+++ clients/admin-ui/src/pages/management/custom-fields.tsx
@@ -1,13 +1,20 @@
-import { Box, Heading, Text } from "@fidesui/react";
+import {
+  Box,
+  Heading,
+  Switch,
+  Text,
+  useDisclosure,
+  WarningIcon,
+} from "@fidesui/react";
+import ConfirmationModal from "common/ConfirmationModal";
 import { useHasPermission } from "common/Restrict";
 import type { NextPage } from "next";
-import { useMemo } from "react";
-import { Column } from "react-table";
+import { ChangeEvent, useMemo } from "react";
+import { CellProps, Column } from "react-table";
 
 import { useAppSelector } from "~/app/hooks";
 import Layout from "~/features/common/Layout";
 import {
-  EnableCell,
   FieldTypeCell,
   TitleCell,
   WrappedCell,
@@ -20,6 +27,71 @@ import {
   useUpdateCustomFieldDefinitionMutation,
 } from "~/features/plus/plus.slice";
 import { CustomFieldDefinitionWithId, ScopeRegistryEnum } from "~/types/api";
+
+type EnableCellProps<T extends object> = CellProps<T, boolean> & {
+  onToggle: (data: any) => Promise<any>;
+};
+
+const EnableCell = <T extends object>({
+  value,
+  column,
+  row,
+  onToggle,
+}: EnableCellProps<T>) => {
+  const modal = useDisclosure();
+  const handlePatch = async ({ enable }: { enable: boolean }) => {
+    // @ts-ignore
+    await onToggle({ ...row.original, active: enable });
+  };
+
+  const handleToggle = async (event: ChangeEvent<HTMLInputElement>) => {
+    const { checked } = event.target;
+    if (checked) {
+      await handlePatch({ enable: true });
+    } else {
+      modal.onOpen();
+    }
+  };
+
+  return (
+    <>
+      <Switch
+        colorScheme="complimentary"
+        isChecked={!value}
+        data-testid={`toggle-${column.Header}`}
+        /**
+         * It's difficult to use a custom column in react-table 7 since we'd have to modify
+         * the declaration file. However, that modifies the type globally, so our datamap table
+         * would also have issues. Ignoring the type for now, but should potentially revisit
+         * if we update to react-table 8
+         * https://github.com/DefinitelyTyped/DefinitelyTyped/discussions/59837
+         */
+        // @ts-ignore
+        disabled={column.disabled}
+        onChange={handleToggle}
+      />
+      <ConfirmationModal
+        isOpen={modal.isOpen}
+        onClose={modal.onClose}
+        onConfirm={() => {
+          handlePatch({ enable: false });
+          modal.onClose();
+        }}
+        title="Disable privacy notice"
+        message={
+          <Text color="gray.500">
+            Are you sure you want to disable this privacy notice? Disabling this
+            notice means your users will no longer see this explanation about
+            your data uses which is necessary to ensure compliance.
+          </Text>
+        }
+        continueButtonText="Confirm"
+        isCentered
+        icon={<WarningIcon color="orange.100" />}
+      />
+    </>
+  );
+};
 
 const CustomFields: NextPage = () => {
   const { isLoading } = useGetAllCustomFieldDefinitionsQuery();
@@ -48,7 +120,7 @@ const CustomFields: NextPage = () => {
       { Header: "Locations", accessor: "resource_type", Cell: WrappedCell },
       {
         Header: "Enable",
-        accessor: "disabled",
+        accessor: (row) => !row.active,
         disabled: !userCanUpdate,
         Cell: EnableCell,
         onToggle: updateCustomFieldDefinitionTrigger,
